fix step1 피드백 반영
