--- src/test/java/nextstep/subway/line/domain/SectionTest.java
@@ -1,10 +1,0 @@
-package nextstep.subway.line.domain;
-
-import org.junit.jupiter.api.DisplayName;
-
-@DisplayName("Station 클래스")
-public class SectionTest {
-    public static Sections 빈_섹션_생성() {
-        return new Sections();
-    }
-}
